requirements.txt
requests   
PyYAML
kubernetes
openshift
python-decouple
urllib3
